session id

+++ src/sharding-config.c
@@ -382,7 +382,7 @@ shard_conf_is_shard_table(const char *db, const char *table)
 }
 
 GPtrArray *
-shard_conf_get_fixed_group(GPtrArray *groups, guint32 fixture)
+shard_conf_get_fixed_group(GPtrArray *groups, guint64 fixture)
 {
     int len = g_list_length(shard_conf_all_groups);
     if (len == 0) {
